Add files via upload
Data Collection for SSL Algorithm Pretraining:

We gathered a total of 36,988 unlabeled images comprising cats and dogs, with 18,499 images of each. These images were resized and normalized to have a mean of 0 and a standard deviation of 1. Subsequently, we developed several image transformation functions to enhance the dataset. These transformations include rotation, greyscaling, color jittering, and masking. Each transformation can be applied individually or in combination to augment the dataset further.